ports old admin stuff to new app
