--- SuperLandoAdmin/AdminLib/Sources/APIClient/GenericResponse.swift
@@ -1,12 +1,0 @@
-//
-//  GenericResponse.swift
-//  
-//
-//  Created by Mauricio Cardozo on 02/08/23.
-//
-
-import Foundation
-
-public struct GenericResponse: Codable, Equatable {
-  public let message: String
-}
