+++ app/LandinhoLib/Sources/Categories/Categories.swift
@@ -1,6 +1,6 @@
 //
-//  File.swift
-//  
+//  Categories.swift
+//
 //
 //  Created by Mauricio Cardozo on 12/11/23.
 //
@@ -17,13 +17,15 @@ public struct Categories: Reducer {
   }
 
   public enum Action: Equatable {
-
+    case onAppear
   }
 
   public var body: some ReducerOf<Self> {
     Reduce { state, action in
       switch action {
-
+      case .onAppear:
+        #warning("TODO call API")
+        return .none
       }
     }
   }
@@ -34,10 +36,7 @@ public struct CategoriesView: View {
   public init() {}
   public var body: some View {
     VStack {
-      Image(systemName: "globe")
-        .imageScale(.large)
-        .foregroundStyle(.tint)
-      Text("Categories!")
+      Text("Placeholder for category list")
     }
     .padding()
   }
