--- app/LandinhoLib/Sources/Common/Common.swift
@@ -1,8 +1,0 @@
-//
-//  File.swift
-//  
-//
-//  Created by Mauricio Cardozo on 10/11/23.
-//
-
-import Foundation
