+++ app/LandinhoLib/Sources/Common/Race.swift
@@ -1,0 +1,14 @@
+//
+//  Race.swift
+//
+//
+//  Created by Mauricio Cardozo on 14/11/23.
+//
+
+import Foundation
+
+public struct Race: Codable, Equatable, Identifiable, Hashable {
+  public let id: UUID
+  public let title: String
+  public let events: [RaceEvent]
+}
