+++ app/LandinhoLib/Sources/ScheduleList/ScheduleList.swift
@@ -1,19 +1,19 @@
 //
-//  File.swift
-//  
+//  ScheduleList.swift
+//
 //
 //  Created by Mauricio Cardozo on 12/11/23.
 //
 
+import APIClient
+import Common
 import Foundation
 import ComposableArchitecture
 import SwiftUI
 
 public struct ScheduleList: Reducer {
 
-  public init() {
-    
-  }
+  public init() {}
 
   public struct State: Equatable {
     public init(categoryTag: String?) {
@@ -21,17 +21,28 @@ public struct ScheduleList: Reducer {
     }
 
     let categoryTag: String?
+
+    var racesState = APIClient<Race>.State(endpoint: "next-race")
   }
 
   public enum Action: Equatable {
+    case onAppear
 
+    case racesRequest(APIClient<Race>.Action)
   }
 
   public var body: some ReducerOf<Self> {
     Reduce { state, action in
       switch action {
-
+      case .onAppear:
+        return .send(.racesRequest(.request(.get)))
+      case .racesRequest:
+        return .none
       }
+    }
+
+    Scope(state: \.racesState, action: /Action.racesRequest) {
+      APIClient()
     }
   }
 }
@@ -45,7 +56,23 @@ public struct ScheduleListView: View {
 
   public var body: some View {
     WithViewStore(store, observe: { $0 }) { viewStore in
-      Text("Placeholder")
+      ScrollView {
+        VStack {
+          switch viewStore.racesState.response {
+          case .idle:
+            Text("idle")
+          case .loading:
+            Text("loading")
+          case .reloading(let race), .finished(.success(let race)):
+            Text("finished or reloading - \(race.title)")
+          case .finished(.failure(let error)):
+            APIErrorView(error: error)
+          }
+        }
+      }
+    }
+    .task {
+      store.send(.onAppear)
     }
   }
 }
