+++ app/LandinhoLib/Sources/Settings/Settings.swift
@@ -1,10 +1,11 @@
 //
-//  File.swift
-//  
+//  Settings.swift
+//
 //
 //  Created by Mauricio Cardozo on 12/11/23.
 //
 
+import Admin
 import Foundation
 import ComposableArchitecture
 import SwiftUI
@@ -14,29 +15,40 @@ public struct Settings: Reducer {
 
   public struct State: Equatable {
     public init() {}
+
+    var adminState = Admin.State()
   }
 
   public enum Action: Equatable {
-
+    case admin(Admin.Action)
   }
 
   public var body: some ReducerOf<Self> {
     Reduce { state, action in
       switch action {
-
+      case .admin:
+        return .none
       }
+    }
+
+    Scope(state: \.adminState, action: /Action.admin) {
+      Admin()
     }
   }
 }
 
 public struct SettingsView: View {
-  public init() {}
+  public init(store: StoreOf<Settings>) {
+    self.store = store
+  }
+
+  let store: StoreOf<Settings>
+
   public var body: some View {
-    NavigationStack {
-      Form {
-        Text("")
-      }
-    }
-    .navigationTitle("Settings")
+    AdminView(store: store.scope(
+      state: \.adminState,
+      action: Settings.Action.admin)
+    )
   }
 }
+
